Added integration of directions API for drawing the route

+++ Breda_op_Stap/app/src/main/java/com/example/breda_op_stap/logic/RouteParser.java
@@ -54,8 +54,8 @@ public class RouteParser
             {
                 String name = obj.getJSONObject(i).getString("name");
                 LatLng position = new LatLng(
-                        obj.getJSONObject(i).getJSONObject("coordinates").getInt("lat"),
-                        obj.getJSONObject(i).getJSONObject("coordinates").getInt("lng")
+                        obj.getJSONObject(i).getJSONObject("coordinates").getDouble("lat"),
+                        obj.getJSONObject(i).getJSONObject("coordinates").getDouble("lng")
                 );
                 String desc = obj.getJSONObject(i).getString("description");
                 ArrayList<String> images = new ArrayList<String>();
